backend/services/news/requirements.txt
numpy
psycopg2-binary
python-dotenv
passlib
fastapi 
uvicorn[standard]
pydantic

httpx
requests
beautifulsoup4